test: add model tests for postgres
