+++ packages/amplify-e2e-tests/src/__tests__/rds-mysql-multi-auth-1.test.ts
@@ -10,23 +10,21 @@ import {
   setupRDSInstanceAndData,
   sleep,
   updateAuthAddUserGroups,
+  getProjectMeta,
 } from 'amplify-category-api-e2e-core';
 import { existsSync, writeFileSync, removeSync } from 'fs-extra';
 import generator from 'generate-password';
 import path from 'path';
 import { schema, sqlCreateStatements } from './auth-test-schemas/userpool-apikey';
 import {
-  createModelOperationHelpers,
   configureAppSyncClients,
   checkOperationResult,
   checkListItemExistence,
   appendAmplifyInput,
   getAppSyncEndpoint,
 } from '../rds-v2-test-utils';
-import { setupUser, getUserPoolId, signInUser, configureAmplify, getConfiguredAppsyncClientCognitoAuth } from '../schema-api-directives';
-import { gql } from 'graphql-tag';
+import { setupUser, getUserPoolId, signInUser, configureAmplify } from '../schema-api-directives';
 import { GQLQueryHelper } from '../query-utils/gql-helper';
-import { check } from 'yargs';
 
 // to deal with bug in cognito-identity-js
 (global as any).fetch = require('node-fetch');
@@ -37,7 +35,7 @@ describe('RDS Cognito userpool provider Auth tests', () => {
   // Generate settings for RDS instance
   const username = db_user;
   const password = db_password;
-  const region = 'ap-northeast-2';
+  let region = 'us-east-1';
   let port = 3306;
   const database = 'default_db';
   let host = 'localhost';
@@ -98,6 +96,9 @@ describe('RDS Cognito userpool provider Auth tests', () => {
       disableAmplifyAppCreation: false,
       name: projName,
     });
+
+    const metaAfterInit = getProjectMeta(projRoot);
+    region = metaAfterInit.providers.awscloudformation.Region;
 
     await addApi(projRoot, {
       transformerVersion: 2,
